pipeline/requirements.txt
anthropic>=0.40.0
supabase>=2.9.0
feedparser>=6.0.11
httpx>=0.27.0
python-dotenv>=1.0.1
resend>=2.5.0
beautifulsoup4>=4.12.3
schedule>=1.2.2
